requirements.txt
pyyaml
click
jinja2
gitpython